tjoin: use long not int for Cartesian axis bins
AbstractCartesianMatchEngine and its concrete subclasses chop the
axes into bins and label each bin with an integer.
In the case of bins that are very small compared to the range of
values on the axis (real case: OBMT values in the range 0-10000
with matching tolerance ~1e-10), this could lead to integer overflow
so that many bins ended up with a value of Integer.MAX_VALUE or
Integer.MIN_VALUE (this value is calcluated using (int) Math.floor(...)).
This is not a correctness error, but either or both of those extremal
bins could end up very full, which would lead to matching so slow it
would probably never terminate.

Improve matters by using longs not ints to label the bins.
This is much less likely to run out of bins; it's still possible
and would again result in a performance issue or likely match failure
rather than a correctness error.

This approach will use more memory for bin keys, but not much,
since there is the overhead in any case of at least one Object
(actually two as currently implemented: the Cell object and the
long[] array it contains) when storing the primitive values.

+++ tjoin/src/main/uk/ac/starlink/table/join/Cell.java
@@ -2,15 +2,15 @@ package uk.ac.starlink.table.join;
 
 /**
  * Represents a cell on a Cartesian grid.
- * This is just a wrapper for an int[] array, with suitable equals and
- * hashCode methods defined.
+ * This is just a wrapper for a <code>long[]</code> array,
+ * with suitable equals and hashCode methods defined.
  *
  * @author   Mark Taylor
  * @since    16 Aug 2011
  */
 class Cell {
 
-    private final int[] label_;
+    private final long[] label_;
 
     /**
      * Constructor.
@@ -18,7 +18,7 @@ class Cell {
      * @param   label  array determining content of this cell;
      *                 it is not cloned
      */
-    public Cell( int[] label ) {
+    public Cell( long[] label ) {
         label_ = label;
     }
 
@@ -26,7 +26,7 @@ class Cell {
     public boolean equals( Object o ) {
         if ( o instanceof Cell ) {
             Cell other = (Cell) o;
-            int[] otherLabel = other.label_;
+            long[] otherLabel = other.label_;
             int ndim = label_.length;
             for ( int i = 0; i < ndim; i++ ) {
                 if ( otherLabel[ i ] != label_[ i ] ) {
@@ -45,7 +45,7 @@ class Cell {
         int code = 37;
         int ndim = label_.length;
         for ( int i = 0; i < ndim; i++ ) {
-            code = 23 * code + label_[ i ];
+            code = 23 * code + (int) label_[ i ];
         }
         return code;
     }
